✨Feat: user-model 작성
# Conflicts:
#	app/core/config.py
