+++ app/core/config.py
@@ -24,16 +24,7 @@ TORTOISE_ORM = {
         "models": {
             "models": [
                 "aerich.models",
-                "app.models.base_user_models",
-                "app.models.corporate_user_models",
-                "app.models.seeker_user_models",
-                "app.models.user_ban_models",
-                "app.models.success_review_models",
-                "app.models.free_board_models",
-                "app.models.resume_models",
-                "app.models.comment_models",
-                "app.models.job_posting_models",
-                "aerich.models",
+                "app.models.user_models",
             ],
             "default_connection": "default",
         }
